+++ lib/diaspora_federation/parsers/xml_parser.rb
@@ -14,7 +14,7 @@ module DiasporaFederation
       def parse(root_node)
         from_xml_sanity_validation(root_node)
 
-        hash = root_node.element_children.map {|child|
+        hash = root_node.element_children.uniq(&:name).map {|child|
           property, type = find_property_for(child.name)
           if property
             value = parse_element_from_node(child.name, type, root_node)