+++ spec/lib/diaspora_federation/entities/account_migration_spec.rb
@@ -8,10 +8,11 @@ module DiasporaFederation
     let(:new_diaspora_id) { new_user.diaspora_id }
 
     let(:data) {
-      hash.tap {|hash|
+      hash.dup.tap {|data|
         properties = described_class.new(hash).send(:enriched_properties)
-        hash[:signature] = properties[:signature]
-        hash[:profile] = Entities::Profile.new(hash[:profile].to_h.tap {|profile| profile[:edited_at] = nil })
+        data[:signature] = properties[:signature]
+        data[:profile] = Entities::Profile.new(hash[:profile].to_h.tap {|profile| profile[:edited_at] = nil })
+        data[:remote_photo_path] = "http://localhost:3000/uploads/images/"
       }
     }
     let(:signature_data) { "AccountMigration:#{old_diaspora_id}:#{new_diaspora_id}" }
@@ -128,6 +129,7 @@ module DiasporaFederation
           </profile>
           <signature>#{data[:signature]}</signature>
           <old_identity>#{data[:old_identity]}</old_identity>
+          <remote_photo_path>#{data[:remote_photo_path]}</remote_photo_path>
         </account_migration>
       XML
 
@@ -167,6 +169,7 @@ module DiasporaFederation
           </profile>
           <signature>#{data[:signature]}</signature>
           <old_identity>#{data[:old_identity]}</old_identity>
+          <remote_photo_path>#{data[:remote_photo_path]}</remote_photo_path>
         </account_migration>
       XML
 
@@ -218,6 +221,51 @@ module DiasporaFederation
           parsed_xml = Nokogiri::XML(xml).root
           Entity.entity_class(parsed_xml.name).from_xml(parsed_xml)
         }.to raise_error Entity::ValidationError
+      end
+    end
+
+    context "optional values" do
+      let(:hash) {
+        {
+          author:  old_diaspora_id,
+          profile: Entities::Profile.new(author: new_diaspora_id)
+        }
+      }
+
+      it "uses default values when parsing" do
+        minimal_xml = <<~XML
+          <account_migration>
+            <author>#{data[:author]}</author>
+            <profile>
+              <author>#{data[:profile].author}</author>
+            </profile>
+            <signature>#{data[:signature]}</signature>
+          </account_migration>
+        XML
+
+        parsed_xml = Nokogiri::XML(minimal_xml).root
+        parsed_instance = Entity.entity_class(parsed_xml.name).from_xml(parsed_xml)
+        expect(parsed_instance.old_identity).to eq(data[:author])
+        expect(parsed_instance.remote_photo_path).to be_nil
+      end
+
+      it "adds old_identity when author is the old identity" do
+        expected_xml = <<~XML
+          <account_migration>
+            <author>#{data[:author]}</author>
+            <profile>
+              <author>#{data[:profile].author}</author>
+              <searchable>true</searchable>
+              <public>false</public>
+              <nsfw>false</nsfw>
+            </profile>
+            <signature>#{data[:signature]}</signature>
+            <old_identity>#{data[:author]}</old_identity>
+          </account_migration>
+        XML
+
+        entity = Entities::AccountMigration.new(hash)
+        expect(entity.to_xml.to_s.strip).to eq(expected_xml.strip)
       end
     end
   end